requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
astroid==2.15.4
async-timeout==4.0.2
attrs==21.4.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.0.1
dill==0.3.6
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
frozenlist==1.3.3
idna==3.4
iniconfig==1.1.1
isort==5.12.0
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.9.0
MarkupSafe==2.0.1
mccabe==0.7.0
multidict==6.0.4
packaging==21.3
platformdirs==3.5.0
pluggy==1.0.0
psycopg2-binary==2.9.6
py==1.11.0
PyJWT==2.7.0
pylint==2.17.4
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.7
pyparsing==3.0.8
pytest==7.1.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
SQLAlchemy==1.4.25
tomli==2.0.1
tomlkit==0.11.8
twilio==8.2.0
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.0.1
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
